1. Web.Core renamed to Web.Infrasructure 2. Most of the endpoints created, validated and tested for TopicController, MessageController, ReportController. 3. User entity now has PasswordHash field.

+++ WebApp.BusinessLogic/Services/UserService.cs
@@ -5,28 +5,37 @@ using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
 using WebApp.BusinessLogic.Validation;
-using WebApp.Core.Entities;
-using WebApp.Core.Interfaces;
-using WebApp.Core.Interfaces.IRepositories;
-using WebApp.Core.Interfaces.IServices;
-using WebApp.Core.Models.TopicModels;
-using WebApp.Core.Models.UserModels;
+using WebApp.Infrastructure.Entities;
+using WebApp.Infrastructure.Interfaces.Auth;
+using WebApp.Infrastructure.Interfaces.IRepositories;
+using WebApp.Infrastructure.Interfaces.IServices;
+using WebApp.Infrastructure.Models.TopicModels;
+using WebApp.Infrastructure.Models.UserModels;
 
 namespace WebApp.BusinessLogic.Services;
+
 public class UserService : IUserService
 {
     private readonly IUnitOfWork unitOfWork;
     private readonly IMapper mapper;
+    private readonly IPasswordHasher passwordHasher;
+    private readonly IJvtProvider jvtProvider;
 
-    public UserService(IUnitOfWork unitOfWork, IMapper mapper)
+    public UserService(
+        IUnitOfWork unitOfWork,
+        IMapper mapper,
+        IPasswordHasher passwordHasher,
+        IJvtProvider jvtProvider)
     {
         this.unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
         this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
+        this.passwordHasher = passwordHasher ?? throw new ArgumentNullException(nameof(passwordHasher));
+        this.jvtProvider = jvtProvider;
     }
 
-    public async Task<IEnumerable<UserModel>> GetAllAsync(TopicQueryParametersModel queryParameters)
+    public async Task<IEnumerable<UserModel>> GetAllAsync(TopicQueryParametersModel? queryParameters)
     {
-        var userEntities = await this.unitOfWork.UserRepository.GetAllAsync();
+        var userEntities = await this.unitOfWork.UserRepository.GetAllWithDetailsAsync();
         var userModels = userEntities.Select(u => this.mapper.MapWithExceptionHandling<UserModel>(u));
 
         return userModels;
@@ -34,27 +43,64 @@ public class UserService : IUserService
 
     public async Task<UserModel> GetByIdAsync(params object[] keys)
     {
+        if (!this.unitOfWork.UserRepository.IsExist((int)keys[0]))
+        {
+            throw new ForumException("User with this Id does not exist");
+        }
+
         var userEntity = await this.unitOfWork.UserRepository.GetByIdAsync(keys);
         var userModel = this.mapper.MapWithExceptionHandling<UserModel>(userEntity);
 
         return userModel;
     }
 
-    public async Task<int> AddAsync(UserCreateModel model)
+    public async Task<int> RegisterAsync(UserRegisterModel model)
     {
-        ForumException.ThrowIfNull(model);
+        try
+        {
+            ForumException.ThrowIfNull(model);
 
-        var user = this.mapper.MapWithExceptionHandling<User>(model);
+            model.PasswordHash = this.passwordHasher.HashPassword(model.PasswordHash);
 
-        int userId = (int) await this.unitOfWork.UserRepository.AddAsync(user);
-        await this.unitOfWork.SaveAsync();
+            var user = this.mapper.MapWithExceptionHandling<User>(model);
 
-        return userId;
+            int userId = (int)await this.unitOfWork.UserRepository.AddAsync(user);
+            await this.unitOfWork.SaveAsync();
+
+            return userId;
+        }
+        catch (Exception)
+        {
+            throw new ForumException("Failed to register user");
+        }
+    }
+
+    public async Task<string> LoginAsync(UserLoginModel model)
+    {
+        var user = await this.unitOfWork.UserRepository.GetByEmailAsync(model.Email);
+
+        if (user is null)
+        {
+            throw new ForumException("Invalid password or email");
+        }
+
+        var result = this.passwordHasher.Verify(model.Password, user.PasswordHash);
+
+        if (result == false)
+        {
+            throw new ForumException("Invalid password or email");
+        }
+
+        var token = this.jvtProvider.GenerateToken(user);
+
+        return token;
     }
 
     public async Task UpdateAsync(UserUpdateModel model)
     {
         ForumException.ThrowIfNull(model);
+
+        var existingUser = this.unitOfWork.UserRepository.GetByIdAsync(model.Id);
 
         var user = this.mapper.MapWithExceptionHandling<User>(model);
         this.unitOfWork.UserRepository.Update(user);
@@ -64,7 +110,26 @@ public class UserService : IUserService
 
     public async Task DeleteAsync(int modelId)
     {
+        if (this.unitOfWork.UserRepository.IsExist(modelId))
+        {
+            throw new ForumException("User with this Id does not exist");
+        }
+
         await this.unitOfWork.UserRepository.DeleteByIdAsync(modelId);
         await this.unitOfWork.SaveAsync();
+    }
+
+    public async Task<UserPublicProfileModel> GetUserByEmailAsync(string email)
+    {
+        var user = await this.unitOfWork.UserRepository.GetByEmailAsync(email);
+
+        if (user == null)
+        {
+            throw new ForumException("User with this Email does not exist");
+        }
+
+        var userProfile = this.mapper.MapWithExceptionHandling<UserPublicProfileModel>(user);
+
+        return userProfile;
     }
 }
